Add files via upload
As for your first question, the CEO has requested a trend of the revenue to see if there is any
seasonality in the store sales. My analysis shows that there are some months of the year where
exceptional growth is witnessed. The data shows that the revenue in the first 8 months is fairly
constant as the average revenue generated for these 8 months is around $685k. The increase
in revenue starts in the month of September, where the revenue increases by 40% over the
previous month. This trend continues till the month of November where it reached 1.5 million
USD, the highest during the entire year. The data is incomplete for the month of December,
therefore, no conclusion can be drawn from it, unfortunately. This analysis shows that the retail
store sales are impacted by the seasonality which usually occurs in the last 4 months of the
year.

The second visual shows how the top 10 countries which have opportunities for growth are
performing. This data does not include the UK as the country already has high demand and I’ve
been told you’re more focused on the countries where demand can be increased. The analysis
shows that countries such as the Netherlands, Ireland, Germany and France have high volumes
of units bought and revenue generated. I would suggest that these countries should be focused
on to ensure that measures are taken to capture these markets even more.

The third analysis has been performed on the top 10 customers who have purchased the most
from the store. The data shows that there is not much of a difference between the purchases
made by the top 10 customers. The highest revenue generating customer only purchased 17%
more than the 2nd highest which shows that the business is not relying only on a few customers
to generate the revenue. This shows that the bargaining power of customers is low and the
business is in a good position.

Finally, the map chart shows the regions that have generated the most revenue compared with
the regions that have not. It can be seen that apart from the UK, countries such as Netherlands,
Ireland, Germany, France and Australia are generating high revenue and the company should
invest more in these areas to increase demand for products. The map also shows that most of
the sales are only in the European region with very few in the American region. Africa and Asia
do not have any demand for the products, along with Russia. A new strategy targeting these
areas has the potential to boost sales revenues and profitability.